reviving

+++ src/main/java/com/rl_todo/Resource.java
@@ -3,9 +3,9 @@ package com.rl_todo;
 public class Resource
 {
     public String myId;
-    public int myAmount;
+    public float myAmount;
 
-    public Resource(String aId, int aAmount)
+    public Resource(String aId, float aAmount)
     {
         myId = aId;
         myAmount = aAmount;
